Fix title and example value of signal/RSSI in flows
The signal strength (e.g. -75 dbM) and RSSI (e.g. 75%) was swapped
around, so % was used to report signal strength, and dBM to report
RSSI, which doesn't match any other places in the app.

This change does not affect any existing automations, as only
the title and example value is updated. The raw flow value
is still the same, with the same key (name).