Use descriptive names for RSSI and signal strength tokens
The device properties are already reporting measurements with
these names, and the documentation also seems to indicate we
should use descriptive names for the token title, instead of
its unit.

https://apps.developer.homey.app/the-basics/flow/tokens

This change does not affect any existing automations, as only
the title is updated. The raw flow value is still the same,
with the same key (name).